Switch out SecondaryNav to PrimaryNav
