+++ src/components/Layout/Layout.test.js
@@ -14,11 +14,6 @@ describe("Layout", () => {
     expect(wrapper.find(".l-side")).toHaveLength(1);
   });
 
-  it("renders without a sidebar if false bool is passed", () => {
-    const wrapper = shallow(<Layout sidebar={false} />);
-    expect(wrapper.find(".l-side")).toHaveLength(0);
-  });
-
   it("should display the children", () => {
     const wrapper = shallow(<Layout>content</Layout>);
     expect(wrapper.find("#main-content").html()).toStrictEqual(
