+++ src/components/PrimaryNav/PrimaryNav.js
@@ -1,72 +1,36 @@
 import React from "react";
 import { Link } from "react-router-dom";
+import classNames from "classnames";
 
-function PrimaryNav() {
+import "./_primary-nav.scss";
+
+const pages = [
+  { label: "Models", path: "/" },
+  { label: "Controllers", path: "/controllers" },
+  { label: "Usage", path: "/usage" },
+  { label: "Logs", path: "/logs" }
+];
+
+const PrimaryNav = () => {
+  const currentLocation = window.location.pathname;
   return (
-    <div>
-      <header id="navigation" className="p-navigation">
-        <div className="p-navigation__banner">
-          <div className="p-navigation__logo">
-            <a className="p-navigation__link" href="/">
-              <img
-                className="p-navigation__image"
-                src="https://assets.ubuntu.com/v1/a9e0ed4a-jaas-logo1.svg"
-                alt="JAAS logo"
-                width="95"
-              />
-            </a>
-          </div>
-          <a
-            href="#navigation"
-            className="p-navigation__toggle--open"
-            title="menu"
+    <nav className="p-primary-nav">
+      <ul className="p-list">
+        {Object.values(pages).map(navItem => (
+          <li
+            key={navItem.path}
+            className={classNames("p-list__item", {
+              "is-selected": currentLocation === navItem.path
+            })}
           >
-            Menu
-          </a>
-          <a
-            href="#navigation-closed"
-            className="p-navigation__toggle--close"
-            title="close menu"
-          >
-            Close menu
-          </a>
-        </div>
-        <nav className="p-navigation__nav">
-          <span className="u-off-screen">
-            <a href="#main-content">Jump to main content</a>
-          </span>
-          <ul className="p-navigation__links" role="menu">
-            <li className="p-navigation__link" role="menuitem">
-              <a className="p-link--external" href="https://jaas.ai/jaas">
-                About
-              </a>
-            </li>
-            <li className="p-navigation__link" role="menuitem">
-              <a
-                className="p-link--external"
-                href="https://discourse.jujucharms.com/"
-              >
-                Discourse
-              </a>
-            </li>
-            <li className="p-navigation__link" role="menuitem">
-              <a className="p-link--external" href="https://jaas.ai/docs">
-                Docs
-              </a>
-            </li>
-            <li className="p-navigation__link" role="menuitem">
-              <a className="p-link--external" href="#_">
-                Charmhub
-              </a>
-            </li>
-            <li className="p-navigation__link" role="menuitem">
-              <Link to="/">Settings</Link>
-            </li>
-          </ul>
-        </nav>
-      </header>
-    </div>
+            <Link className="p-list__link" to={navItem.path}>
+              {navItem.label}
+            </Link>
+          </li>
+        ))}
+      </ul>
+    </nav>
   );
-}
+};
 
 export default PrimaryNav;
