+++ src/components/PrimaryNav/_primary-nav.scss
@@ -1,6 +1,6 @@
 @import "vanilla-framework/scss/settings_colors";
 
-.p-secondary-nav {
+.p-primary-nav {
   .p-list__link {
     color: $color-dark;
   }
